requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=2.2.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
google-generativeai>=0.3.0
requests>=2.31.0
langchain-experimental>=0.0.49
langchain>=0.1.0
langchain-community>=0.0.21
langchain-google-genai>=0.0.6
ollama>=0.1.6
tabulate
scikit-learn